+++ lib/src/events.dart
@@ -228,6 +228,17 @@ class LocalTrackPublishedEvent with RoomEvent, ParticipantEvent {
       '(participant: ${participant}, publication: ${publication})';
 }
 
+class LocalTrackSubscribedEvent with RoomEvent, ParticipantEvent {
+  final String trackSid;
+  const LocalTrackSubscribedEvent({
+    required this.trackSid,
+  });
+
+  @override
+  String toString() => '${runtimeType}'
+      '(trakSid: ${trackSid}})';
+}
+
 /// The local participant has unpublished one of their [Track].
 /// Emitted by [Room] and [LocalParticipant].
 class LocalTrackUnpublishedEvent with RoomEvent, ParticipantEvent {
@@ -451,15 +462,15 @@ class ParticipantPermissionsUpdatedEvent with RoomEvent, ParticipantEvent {
 class TranscriptionSegment {
   final String id;
   final String text;
-  final DateTime startTime;
-  final DateTime endTime;
+  final DateTime firstReceivedTime;
+  final DateTime lastReceivedTime;
   final bool isFinal;
   final String language;
   const TranscriptionSegment({
     required this.id,
     required this.text,
-    required this.startTime,
-    required this.endTime,
+    required this.firstReceivedTime,
+    required this.lastReceivedTime,
     required this.isFinal,
     required this.language,
   });