+++ lib/src/internal/events.dart
@@ -170,6 +170,14 @@ class EngineDisconnectedEvent with InternalEvent, EngineEvent {
 }
 
 @internal
+class EngineLocalTrackSubscribedEvent with InternalEvent, EngineEvent {
+  final String trackSid;
+  const EngineLocalTrackSubscribedEvent({
+    required this.trackSid,
+  });
+}
+
+@internal
 class EngineFullRestartingEvent with InternalEvent, EngineEvent {
   const EngineFullRestartingEvent();
 }
@@ -298,6 +306,15 @@ class SignalTrackUnpublishedEvent with SignalEvent, InternalEvent {
 }
 
 @internal
+class SignalLocalTrackSubscribedEvent with SignalEvent, InternalEvent {
+  final String trackSid;
+
+  const SignalLocalTrackSubscribedEvent({
+    required this.trackSid,
+  });
+}
+
+@internal
 class SignalRoomUpdateEvent with SignalEvent, InternalEvent {
   final lk_models.Room room;
 
@@ -326,11 +343,14 @@ class EngineActiveSpeakersUpdateEvent with EngineEvent, InternalEvent {
 
 @internal
 class SignalLeaveEvent with SignalEvent, InternalEvent {
-  final bool canReconnect;
-  final lk_models.DisconnectReason reason;
+  bool get canReconnect => request.canReconnect;
+  lk_rtc.LeaveRequest_Action get action => request.action;
+  lk_models.DisconnectReason get reason => request.reason;
+  lk_rtc.RegionSettings? get regions =>
+      request.hasReason() ? request.regions : null;
+  final lk_rtc.LeaveRequest request;
   const SignalLeaveEvent({
-    required this.canReconnect,
-    required this.reason,
+    required this.request,
   });
 }