+++ lib/src/livekit.dart
@@ -12,8 +12,24 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+import 'package:flutter_webrtc/flutter_webrtc.dart' as rtc;
+
+import 'package:livekit_client/livekit_client.dart';
+
 /// Main entry point to connect to a room.
 /// {@category Room}
 class LiveKitClient {
-  static const version = '2.2.4';
+  static const version = '2.3.0';
+
+  /// Initialize the WebRTC plugin. If this is not manually called, will be
+  /// initialized with default settings.
+  /// This method must be called before calling any LiveKit SDK API.
+  static Future<void> initialize({bool bypassVoiceProcessing = false}) async {
+    if (lkPlatformIsMobile()) {
+      await rtc.WebRTC.initialize(options: {
+        if (bypassVoiceProcessing)
+          'bypassVoiceProcessing': bypassVoiceProcessing,
+      });
+    }
+  }
 }