+++ CHANGELOG.md
@@ -1,5 +1,35 @@
 # CHANGELOG
 
+## 2.3.0
+
+* fix: Fixed speaker cannot be switched on iOS. (#617)
+* feat: Increase default audio bitrate. (#616)
+
+## 2.2.8
+
+* Fix TrackStreamStateUpdatedEvent not emitted. (#612)
+
+## 2.2.7
+
+* feat: support bypass voice processing. (#595)
+* fix: Dispose LocalParticipant when room.disconnect. (#609)
+* Fix spelling error in method name: _checkPremissions to _checkPermissions (#605)
+* fix: fix speaker switching behavior for android. (#604)
+* fix: Optimize stats display (#602)
+* fix: Handle disconnect reasons correctly. (#594)
+
+## 2.2.6
+
+* fix: android build failure. (#587)
+* Update README.md for Android Screenshare (#583)
+
+## 2.2.5
+
+* upgrade flutter-webrtc to 0.11.7
+* upgrade web to 1.0.0
+* feat: Add timestamps to TranscriptionSegment. (#577)
+* feat: Prepare connection/region pinning. (#574)
+
 ## 2.2.4
 
 * fix bug for PlatformView on iOS (#570)